Store error messages in separate file
